Changes for V.4 and ANSI C
git-svn-id: https://svn.so-much-stuff.com/svn/trunk/cvs/trunk@1448 93276802-5317-4dee-a4ef-7ae40ebda4b9

+++ games.d/hack1.0.3.d/hack.rip.c
@@ -42,7 +42,7 @@ outrip(){
 		strchr(vowels, *killer) ? " an" : " a");
 	center(8, buf);
 	(void) strcpy(buf, killer);
-	if(strlen(buf) > 16) {
+	if(strlen(buf) > (unsigned)16) {
 	    register int i,i0,i1;
 		i0 = i1 = 0;
 		for(i = 0; i <= 16; i++)
@@ -76,6 +76,6 @@ outrip(){
 center(line, text) int line; char *text; {
 register char *ip,*op;
 	ip = text;
-	op = &rip[line][28 - ((strlen(text)+1)/2)];
+	op = &rip[line][28 - ((strlen(text)+1)/(unsigned)2)];
 	while(*ip) *op++ = *ip++;
 }
